chore(dashboards): remove starter-dashboard-v2 experiment code
The starter-dashboard-v2 experiment is over and the test variant is live.
Always create the refreshed "Your starter dashboard" for new projects and
drop the feature-flag check, exposure event, and control-arm template
selection.

Generated-By: PostHog Code
Task-Id: ad13b6fb-fc95-4409-8389-518498304a9c
