+++ products/dashboards/backend/models/dashboard_templates.py
@@ -66,8 +66,8 @@ class DashboardTemplate(UUIDTModel, RootTeamMixin):
         """
         The authoritative hardcoded "Product analytics" template the rest of the codebase relies on
         (the DEFAULT_APP seed, the DB-absent fallback in create_dashboard_from_template, and the global
-        template seed used in tests). It is intentionally the legacy shape; the starter-dashboard-v2
-        test arm opts into the refreshed layout via default_signup_template().
+        template seed used in tests). It is intentionally the legacy shape; new projects get the
+        refreshed layout via default_signup_template().
 
         This OG template is not stored in https://github.com/PostHog/templates-repository
         The system assumes this template is always present and doesn't wait to import it from the template repository
@@ -77,7 +77,7 @@ class DashboardTemplate(UUIDTModel, RootTeamMixin):
     @staticmethod
     def default_signup_template() -> "DashboardTemplate":
         """
-        Refreshed signup dashboard (test arm of the starter-dashboard-v2 experiment).
+        Refreshed signup dashboard given to every new non-demo project's primary dashboard.
         This OG template is not stored in https://github.com/PostHog/templates-repository
         The system assumes this template is always present and doesn't wait to import it from the template repository
         """
@@ -511,7 +511,7 @@ class DashboardTemplate(UUIDTModel, RootTeamMixin):
     @staticmethod
     def legacy_signup_template() -> "DashboardTemplate":
         """
-        Pre-2026 default signup dashboard (control arm of starter-dashboard-v2 experiment).
+        The legacy "Product analytics" dashboard, still used by the DEFAULT_APP template seed.
         This OG template is not stored in https://github.com/PostHog/templates-repository
         The system assumes this template is always present and doesn't wait to import it from the template repository
         """
